% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_qi.R
\name{sim_qi}
\alias{sim_qi}
\title{Get simulations from a model object}
\usage{
sim_qi(mod, nsim = 1000, newdata)
}
\arguments{
\item{mod}{a model object}

\item{nsim}{number of simulations to be run, defaults to 1,000}

\item{newdata}{A data frame with a hypothetical prediction grid.
If absent, defaults to the model frame.}
}
\value{
\code{sim_qi()()} returns a data frame (as a \code{tibble}) with the
quantities of interest and identifying information about the particular
simulation number.
}
\description{
\code{sim_qi()} is a function to simulate quantities of interest
from a regression model
}
\examples{

set.seed(8675309)

M1 <- lm(mpg ~ cyl + wt, mtcars)

sim_qi(M1, 10)


}
